backend/requirements.txt
Django>=5.0.6
djangorestframework>=3.15.1
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
google-auth>=2.30.0
anthropic>=0.40.0
requests>=2.32.3
python-dotenv>=1.0.1
Pillow>=11.0.0
